+++ match_event_templ.go
@@ -225,7 +225,7 @@ func listMatchEvents(matchState MatchState, matchEvents []MatchEvent) templ.Comp
 	})
 }
 
-func addMatchEvent(meta MatchMeta, matchId uint64, isOpen bool) templ.Component {
+func addMatchEvent(meta MatchMeta, matchId uint, isOpen bool) templ.Component {
 	return templ.ComponentFunc(func(ctx context.Context, templ_7745c5c3_W io.Writer) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_Buffer, templ_7745c5c3_IsBuffer := templ_7745c5c3_W.(*bytes.Buffer)
 		if !templ_7745c5c3_IsBuffer {
@@ -496,7 +496,7 @@ func addMatchEvent(meta MatchMeta, matchId uint64, isOpen bool) templ.Component 
 	})
 }
 
-func editMatchEvent(meta MatchMeta, event MatchEvent, isOpen bool, matchId uint64) templ.Component {
+func editMatchEvent(meta MatchMeta, event MatchEvent, isOpen bool, matchId uint) templ.Component {
 	return templ.ComponentFunc(func(ctx context.Context, templ_7745c5c3_W io.Writer) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_Buffer, templ_7745c5c3_IsBuffer := templ_7745c5c3_W.(*bytes.Buffer)
 		if !templ_7745c5c3_IsBuffer {
