+++ player_templ.go
@@ -107,7 +107,7 @@ func playerName(player Player) templ.Component {
 	})
 }
 
-func playerInputSelector(players []Player, playerId uint64, inputType string) templ.Component {
+func playerInputSelector(players []Player, playerId uint, inputType string) templ.Component {
 	return templ.ComponentFunc(func(ctx context.Context, templ_7745c5c3_W io.Writer) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_Buffer, templ_7745c5c3_IsBuffer := templ_7745c5c3_W.(*bytes.Buffer)
 		if !templ_7745c5c3_IsBuffer {
@@ -1118,7 +1118,7 @@ func viewPlayerPayments(playerPayments []PlayerPaymentsWithTotals, players []Pla
 	})
 }
 
-func addPlayerPayment(players []Player, seasonId uint64) templ.Component {
+func addPlayerPayment(players []Player, seasonId uint) templ.Component {
 	return templ.ComponentFunc(func(ctx context.Context, templ_7745c5c3_W io.Writer) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_Buffer, templ_7745c5c3_IsBuffer := templ_7745c5c3_W.(*bytes.Buffer)
 		if !templ_7745c5c3_IsBuffer {
